Moved the parser to a MessageParser class and fixed some setup problems
