+++ app/src/main/java/mindcar/testing/MainActivity.java
@@ -88,87 +88,13 @@ public class MainActivity extends Activity {
     private final Handler handler = new Handler() {
         @Override
         public void handleMessage(Message msg) {
-            Connected.write(checkCommand(car));
-            switch (msg.what) {
-                case TGDevice.MSG_STATE_CHANGE:
-
-                    switch (msg.arg1) {
-                        case TGDevice.STATE_IDLE:
-                            break;
-                        case TGDevice.STATE_CONNECTING:
-                            tv.append("Connecting...\n");
-                            break;
-                        case TGDevice.STATE_CONNECTED:
-                            tv.append("Connected.\n");
-                            tgDevice.start();
-                            break;
-                        case TGDevice.STATE_NOT_FOUND:
-                            tv.append("Can't find\n");
-                            break;
-                        case TGDevice.STATE_NOT_PAIRED:
-                            tv.append("not paired\n");
-                            break;
-                        case TGDevice.STATE_DISCONNECTED:
-                            tv.append("Disconnected mang\n");
-                    }
-
-                    break;
-                case TGDevice.MSG_POOR_SIGNAL:
-                    //signal = msg.arg1;
-                    String tmpNoise = "Noise: " + msg.arg1 + "\n";
-                    noise.setText(tmpNoise);
-                    break;
-                case TGDevice.MSG_RAW_DATA:
-                    //raw1 = msg.arg1;
-                    String tmpRaw = "Raw Data: " + msg.arg1 + "\n";
-                    raw.setText(tmpRaw);
-                    break;
-                case TGDevice.MSG_HEART_RATE:
-                    tv.append("Heart rate: " + msg.arg1 + "\n");
-                    break;
-                case TGDevice.MSG_ATTENTION:
-                    //att = msg.arg1;
-                    eeg.setAttention(msg.arg1);
-                    String tmpAtt = "Attention: " + msg.arg1 + "\n";
-                    att.setText(tmpAtt);
-                    
-                    //Log.v("HelloA", "Attention: " + att + "\n");
-                    break;
-                case TGDevice.MSG_MEDITATION:
-                    eeg.setMeditation(msg.arg1);
-                    String tmpMed = "Meditation: " + msg.arg1 + "\n";
-                    med.setText(tmpMed);
-                    break;
-                case TGDevice.MSG_BLINK:
-                    eeg.setBlink(msg.arg1);
-                    String tmpBlink = "Blink: " + msg.arg1 + "\n";
-                    blink.setText(tmpBlink);
-                    break;
-                case TGDevice.MSG_RAW_COUNT:
-                    //tv.append("Raw Count: " + msg.arg1 + "\n");
-                    break;
-                case TGDevice.MSG_LOW_BATTERY:
-                    Toast.makeText(getApplicationContext(), "Low battery!", Toast.LENGTH_SHORT).show();
-                    break;
-                case TGDevice.MSG_RAW_MULTI:
-                   // TGRawMulti rawM = (TGRawMulti)msg.obj;
-                    //String tmpRaw = "Raw1: " + rawM.ch1 + "\nRaw2: " + rawM.ch2;
-                    //raw.setText(tmpRaw);
-                    break;
-                default:
-                    break;
-            }
-            car.setCommand(eeg);
+            Connected.write(CommandUtils.toByteArray(car.getCommands()));
+            MessageParser.parseMessage(msg, car, eeg);
         }
     };
 
     public void doStuff(View view) {
         if(tgDevice.getState() != TGDevice.STATE_CONNECTING && tgDevice.getState() != TGDevice.STATE_CONNECTED)
             tgDevice.connect(rawEnabled);
-    }
-    
-    public byte[] checkCommand(SmartCar car){
-        return CommandUtils.toByteArray(car);   
-    }
     }
 }
